+++ processor/src/test/resources/expected/com/example/repository/Arez_InjectDisabledRepository.java
@@ -120,6 +120,7 @@ final class Arez_InjectDisabledRepository extends InjectDisabledRepository imple
     if ( Arez.shouldCheckApiInvariants() ) {
       Guards.apiInvariant( () -> ComponentState.isActive( this.$$arezi$$_state ), () -> "Method named 'setName' invoked on " + ComponentState.describe( this.$$arezi$$_state ) + " component named '" + $$arezi$$_name() + "'" );
     }
+    assert null != name;
     if ( !Objects.equals( name, super.getName() ) ) {
       this.$$arez$$_name.preReportChanged();
       super.setName(name);